Fix #71 - Swallow overwrite exception for symlinked files
On some devices the data/data/app_package/lib directory is just a symlink while
on other devices that directory contains the actual extracted libs.
The actual files can be overwritten by us.
The symlinked files (and directory) is only readable by us and so when trying
to create a new FileOutputStream for such a file we would get a
FileNotFoundException.
To avoid crashing in such scenarios we'll swallow the exception if and only if
we can confidently know the file exists, just we don't have write access to it.