For #72 - Enforce unique backup name when creating
Users could previously create a new backup specifying as backup name the title
of an already existing backup, thus overwriting the previous backup.
As a simple solution in the case of a duplicated name we'll append a digit
suffix in the form "(\d+)" to the title specified by the user and display a
Toast informing her about this.