For #72 - Enforce unique backup name when editing
Editing a backup to have the same filename as the name of an already existing
backup will overwrite the existing one.
As a simple solution in the case of a duplicated name we'll append a digit
suffix in the form "(\d+)" to the title specified by the user and display a
Toast informing her about this.